refactor(module:tooltip): flatten base folder
