--- components/tooltip/base/public-api.ts
@@ -1,10 +1,0 @@
-/**
- * @license
- * Copyright Alibaba.com All Rights Reserved.
- *
- * Use of this source code is governed by an MIT-style license that can be
- * found in the LICENSE file at https://github.com/NG-ZORRO/ng-zorro-antd/blob/master/LICENSE
- */
-
-export * from './nz-tooltip-base.component';
-export * from './nz-tooltip-base.directive';
